Refactor macros and unstable functions into submodules
